requirements.txt
dotenv==0.9.9
python-telegram-bot==22.0
